css/css-contain/container-queries/percentage-padding-orthogonal.html

<!doctype html>
<title>CSS Container Queries Test: @container queries affecting height affecting percentage padding</title>
<link rel="help" href="https://drafts.csswg.org/css-contain-3/#container-queries">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/cq-testcommon.js"></script>
<style>
  #vertical {
    width: 500px;
    writing-mode: vertical-lr;
  }

  #padded {
    width: 100%;
    box-sizing: border-box;
    padding-right: 100%;
    background-color: orange;
  }

  #horizontal {
    writing-mode: horizontal-tb;
    width: 100%;
  }

  #container {
    width: 100%;
    container-type: inline-size;
    background-color: green;
  }

  #first, #second { height: 50px; }

  @container size(width = 400px) {
    #second { display: none; }
  }
</style>
<div id="vertical">
  <div id="padded">
    <div id="horizontal">
      <div id="container">
        <div id="first"></div>
        <div id="second"></div>
      </div>
    </div>
  </div>
</div>
<script>
  setup(() => assert_implements_container_queries());

  test(() => assert_equals(container.offsetHeight, padded.offsetHeight),
       "#container height should affect #padded height");
</script>
